requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
pytesseract==0.3.10
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.31.0 